networkUp only after active for 5s
local revert

exception before creating event in db
