+++ code/iaas/api-logic/src/main/java/io/cattle/platform/iaas/api/filter/serviceevent/ServiceEventFilter.java
@@ -3,12 +3,16 @@ package io.cattle.platform.iaas.api.filter.serviceevent;
 
 import io.cattle.platform.api.auth.Policy;
 import io.cattle.platform.api.utils.ApiUtils;
+import io.cattle.platform.archaius.util.ArchaiusUtil;
 import io.cattle.platform.core.constants.AgentConstants;
+import io.cattle.platform.core.constants.CommonStatesConstants;
+import io.cattle.platform.core.constants.ServiceConstants;
 import io.cattle.platform.core.dao.AgentDao;
 import io.cattle.platform.core.dao.ServiceDao;
 import io.cattle.platform.core.model.Agent;
 import io.cattle.platform.core.model.HealthcheckInstance;
 import io.cattle.platform.core.model.HealthcheckInstanceHostMap;
+import io.cattle.platform.core.model.Service;
 import io.cattle.platform.core.model.ServiceEvent;
 import io.cattle.platform.iaas.api.filter.common.AbstractDefaultResourceManagerFilter;
 import io.cattle.platform.object.ObjectManager;
@@ -19,9 +23,16 @@ import io.github.ibuildthecloud.gdapi.request.ApiRequest;
 import io.github.ibuildthecloud.gdapi.request.resource.ResourceManager;
 import io.github.ibuildthecloud.gdapi.util.ResponseCodes;
 
+import static io.cattle.platform.core.model.tables.ServiceTable.SERVICE;
+
+import java.util.Date;
+import java.util.List;
+
 import javax.inject.Inject;
 
 public class ServiceEventFilter extends AbstractDefaultResourceManagerFilter {
+
+    private static final long STAY_ACTIVE_MILLISECONDS = ArchaiusUtil.getLong("stay_active.milliseconds").getValue();
 
     public static final String VERIFY_AGENT = "CantVerifyHealthcheck";
 
@@ -58,6 +69,7 @@ public class ServiceEventFilter extends AbstractDefaultResourceManagerFilter {
             throw new ClientVisibleException(ResponseCodes.FORBIDDEN, VERIFY_AGENT);
         }
 
+
         HealthcheckInstanceHostMap healthcheckInstanceHostMap = null;
         String[] splitted = event.getHealthcheckUuid().split("_");
         if (splitted.length > 2) {
@@ -85,10 +97,39 @@ public class ServiceEventFilter extends AbstractDefaultResourceManagerFilter {
             throw new ClientVisibleException(ResponseCodes.FORBIDDEN, VERIFY_AGENT);
         }
 
+        if (!isNetworkUp(resourceAccId)) {
+            throw new ClientVisibleException(ResponseCodes.CONFLICT);
+        }
+
         event.setInstanceId(healthcheckInstance.getInstanceId());
         event.setHealthcheckInstanceId(healthcheckInstance.getId());
         event.setHostId(healthcheckInstanceHostMap.getHostId());
 
         return super.create(type, request, next);
+    }
+
+    private boolean isNetworkUp(long accountId) {
+        Service networkDriverService = objectManager.findAny(Service.class, SERVICE.ACCOUNT_ID, accountId, SERVICE.REMOVED, null, SERVICE.KIND,
+                ServiceConstants.KIND_NETWORK_DRIVER_SERVICE);
+        if (networkDriverService == null) {
+            return true;
+        }
+        List<Service> services = objectManager.find(Service.class, SERVICE.ACCOUNT_ID, accountId, SERVICE.REMOVED, null, SERVICE.STACK_ID,
+                networkDriverService.getStackId());
+        for (Service service : services) {
+            if (!service.getState().equals(CommonStatesConstants.ACTIVE)) {
+                return false;
+            }
+            Object obj = DataAccessor.fields(service).withKey(ServiceConstants.LAST_ACTIVE).get();
+            if(obj != null) {
+                long lastActiveTime = (long) obj;
+                long currentTime = new Date().getTime();
+                // networkUp only if service has been active for STAY_ACTIVE_SECONDS
+                if (lastActiveTime + STAY_ACTIVE_MILLISECONDS > currentTime) {
+                    return false;
+                }
+            }
+        }
+        return true;
     }
 }
