+++ code/iaas/healthcheck/src/main/java/io/cattle/iaas/healthcheck/process/ServiceEventCreate.java
@@ -3,12 +3,9 @@ package io.cattle.iaas.healthcheck.process;
 import static io.cattle.platform.core.model.tables.ServiceTable.*;
 
 import io.cattle.iaas.healthcheck.service.HealthcheckService;
-import io.cattle.platform.core.constants.CommonStatesConstants;
 import io.cattle.platform.core.constants.HealthcheckConstants;
-import io.cattle.platform.core.constants.ServiceConstants;
 import io.cattle.platform.core.dao.ServiceDao;
 import io.cattle.platform.core.model.HealthcheckInstanceHostMap;
-import io.cattle.platform.core.model.Service;
 import io.cattle.platform.core.model.ServiceEvent;
 import io.cattle.platform.engine.handler.HandlerResult;
 import io.cattle.platform.engine.handler.ProcessHandler;
@@ -17,20 +14,14 @@ import io.cattle.platform.engine.process.ProcessState;
 import io.cattle.platform.process.common.handler.AbstractObjectProcessHandler;
 import io.cattle.platform.util.type.Priority;
 
-import io.github.ibuildthecloud.gdapi.exception.ClientVisibleException;
-import io.github.ibuildthecloud.gdapi.util.ResponseCodes;
-
-import java.util.Arrays;
-import java.util.List;
-
 import javax.inject.Inject;
 import javax.inject.Named;
+
 
 
 @Named
 public class ServiceEventCreate extends AbstractObjectProcessHandler implements ProcessHandler, Priority {
 
-    private static final List<String> UP_STATES = Arrays.asList(CommonStatesConstants.ACTIVE, CommonStatesConstants.UPDATING_ACTIVE);
     @Inject
     HealthcheckService healthcheckService;
 
@@ -59,12 +50,7 @@ public class ServiceEventCreate extends AbstractObjectProcessHandler implements 
         if ("INIT".equals(event.getReportedHealth())) {
             return null;
         }
-
-        if (isNetworkUp(event.getAccountId())) {
-            processHealthcheck(event);
-        } else {
-            throw new ClientVisibleException(ResponseCodes.CONFLICT);
-        }
+        processHealthcheck(event);
 
         return null;
     }
@@ -86,26 +72,6 @@ public class ServiceEventCreate extends AbstractObjectProcessHandler implements 
             healthcheckService.updateHealthcheck(uuid, event.getExternalTimestamp(),
                     getHealthState(event.getReportedHealth()));
         }
-    }
-
-    private boolean isNetworkUp(long accountId) {
-        Service networkDriverService = objectManager.findAny(Service.class, SERVICE.ACCOUNT_ID, accountId, SERVICE.REMOVED, null, SERVICE.KIND,
-                ServiceConstants.KIND_NETWORK_DRIVER_SERVICE);
-        if (networkDriverService == null) {
-            return true;
-        }
-        if (!UP_STATES.contains(networkDriverService.getState())) {
-            return false;
-        }
-        List<Service> services = objectManager.find(Service.class, SERVICE.ACCOUNT_ID, accountId, SERVICE.REMOVED, null, SERVICE.STACK_ID,
-                networkDriverService.getStackId());
-        for (Service service : services) {
-            if (!UP_STATES.contains(service.getState())) {
-                return false;
-            }
-        }
-
-        return true;
     }
 
     protected String getHealthState(String reportedHealth) {
