+++ code/packaging/app-config/src/main/resources/META-INF/cattle/core-model/defaults.properties
@@ -17,6 +17,9 @@ object.link.ignore.serviceconsumemaprecord.accountid=true
 object.link.ignore.serviceexposehostmaprecord.accountid=true
 object.link.name.hostlabelmaprecord.labelid.override=hostLabels
 
+# duration network service needs to stay active for before network's considered Up
+stay_active.milliseconds=5000
+
 # max rows to delete in a single transaction
 cleanup.query_limit.rows=1000
